reorganize modules

+++ bootstrap.sh
@@ -1,4 +1,19 @@
 #!/bin/sh
+
+# validate that requisite submodules are present
+if [ ! -r src/dfxml/.git ] ;
+then
+  echo bringing in submodules
+  echo next time check out with git clone --recursive
+  git submodule init
+fi
+
+if [ ! -r src/Boost-Btree-64-bit-port/.git ] ;
+then
+  echo submodule error.  Please check out with git clone --recursive
+  exit 1
+fi
+
 /bin/rm -rf aclocal.m4
 autoheader -f
 aclocal -I m4
@@ -14,16 +29,4 @@ fi
 
 autoconf -f
 automake --add-missing --copy
-
-# the git dfxml module is required in src/ path
-if [ ! -r src/dfxml ] ;
-then
-  # until hashdb is on git, the dfxml module must be downloaded separately
-  echo bringing in the dfxml submodule from git...
-  cd src; git clone https://github.com/simsong/dfxml.git; cd ..
-
-#  echo bringing in submodules
-#  echo next time check out with git clone --recursive
-#  git submodule init
-fi
